requirements.txt
requests>=2.32.4
pytest>=6.2.5
pytest-html>=3.1.1
python-dotenv>=0.19.0
allure-pytest>=2.9.45
pytest-xdist>=3.0.0
allure-python-commons~=2.15.0